requirements.txt
fastapi==0.104.1
hypercorn==0.15.0
dnspython==2.4.2
email-validator==2.1.0.post1
pylint==2.13.4
pytest==7.4.3
coverage==7.3.2
pytest-cov==4.1.0
pytest-asyncio==0.23.2
uvicorn==0.24.0.post1
certifi==2023.11.17
charset-normalizer==3.3.2
requests==2.31.0
urllib3==2.1.0
prometheus-client==0.19.0
httpx==0.25.2
# Library for MySQL database connection
mysql-connector-python==8.2.0
# Library for model validation
pydantic==2.5.2
